msm8956: Fix camera build

+++ camera/QCamera2/HAL3/QCamera3Channel.cpp
@@ -34,6 +34,7 @@
 #include <cstdlib>
 #include <stdio.h>
 #include <string.h>
+#include <sys/stat.h>
 #include <hardware/camera3.h>
 #include <system/camera_metadata.h>
 #include <gralloc_priv.h>
@@ -524,7 +525,7 @@ void QCamera3Channel::dumpYUV(mm_camera_buf_def_t *frame, cam_dimension_t dim,
                         }
                     }
                     CDBG_HIGH("%s: written number of bytes %ld\n",
-                             __func__, written_len);
+                             __func__, (long)written_len);
                     dumpFrmCnt++;
                     close(file_fd);
                 } else {
@@ -1264,7 +1265,7 @@ void QCamera3RawChannel::dumpRawSnapshot(mm_camera_buf_def_t *frame)
        int file_fd = open(buf, O_RDWR| O_CREAT, 0644);
        if (file_fd >= 0) {
           ssize_t written_len = write(file_fd, frame->buffer, frame->frame_len);
-          ALOGE("%s: written number of bytes %zd", __func__, written_len);
+          ALOGE("%s: written number of bytes %ld", __func__, written_len);
           close(file_fd);
        } else {
           ALOGE("%s: failed to open file to dump image", __func__);
@@ -1461,7 +1462,7 @@ void QCamera3RawDumpChannel::dumpRawSnapshot(mm_camera_buf_def_t *frame)
             if (file_fd >= 0) {
                 ssize_t written_len =
                         write(file_fd, frame->buffer, offset.frame_len);
-                CDBG("%s: written number of bytes %zd", __func__, written_len);
+                CDBG("%s: written number of bytes %ld", __func__, written_len);
                 close(file_fd);
             } else {
                 ALOGE("%s: failed to open file to dump image", __func__);
